@@ -10,25 +10,25 @@
     "@mui/icons-material": "^5.5.1",
     "@mui/material": "^5.5.2",
     "@mui/styled-engine": "^5.5.2",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.26.1",
-    "chart.js": "3.4.1",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "axios": "^1.5.1",
+    "chart.js": "^4.4.0",
     "chroma-js": "2.4.2",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-chartjs-2": "3.0.4",
-    "react-dom": "^17.0.2",
-    "react-github-btn": "1.2.1",
-    "react-meta-tags": "^1.0.1",
-    "react-router-dom": "6.2.1",
-    "react-scripts": "5.0.0",
-    "react-table": "7.7.0",
-    "stylis": "4.0.13",
+    "react": "^18.2.0",
+    "react-chartjs-2": "^5.2.0",
+    "react-dom": "^18.2.0",
+    "react-github-btn": "^1.4.0",
+    "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.16.0",
+    "react-scripts": "^5.0.1",
+    "react-table": "^7.8.0",
+    "stylis": "^4.3.0",
     "stylis-plugin-rtl": "2.1.1",
-    "web-vitals": "2.1.4",
-    "yup": "0.32.11"
+    "web-vitals": "^3.4.0",
+    "yup": "^1.3.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -58,12 +58,12 @@
   "devDependencies": {
     "eslint": "^8.12.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "2.5.1"
+    "prettier": "^3.0.3"
   }
 }
