@@ -14,7 +14,7 @@ Coded by www.creative-tim.com
 */
 
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "App";
 import { AuthContextProvider } from "context";
@@ -22,13 +22,15 @@ import { AuthContextProvider } from "context";
 // Material Dashboard 2 React Context  Provider
 import { MaterialUIControllerProvider } from "context";
 
-ReactDOM.render(
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
+
+root.render(
   <BrowserRouter>
     <AuthContextProvider>
       <MaterialUIControllerProvider>
         <App />
       </MaterialUIControllerProvider>
     </AuthContextProvider>
-  </BrowserRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 );
