@@ -19,7 +19,8 @@ import { useMemo } from "react";
 import PropTypes from "prop-types";
 
 // react-chartjs-2 components
-import { Bar } from "react-chartjs-2";
+import 'chart.js/auto';
+import { Chart } from 'react-chartjs-2';
 
 // @mui material components
 import Card from "@mui/material/Card";
@@ -51,7 +52,7 @@ function ReportsBarChart({ color, title, description, date, chart }) {
               mt={-5}
               height="12.5rem"
             >
-              <Bar data={data} options={options} />
+              <Chart type="bar" data={data} options={options} />
             </MDBox>
           ),
           [chart, color]
