@@ -19,7 +19,8 @@ import { useMemo } from "react";
 import PropTypes from "prop-types";
 
 // react-chartjs-2 components
-import { Doughnut } from "react-chartjs-2";
+import 'chart.js/auto';
+import { Chart } from 'react-chartjs-2';
 
 // @mui material components
 import Card from "@mui/material/Card";
@@ -70,7 +71,7 @@ function DefaultDoughnutChart({ icon, title, description, height, chart }) {
       {useMemo(
         () => (
           <MDBox height={height}>
-            <Doughnut data={data} options={options} />
+            <Chart type="doughnut" data={data} options={options} />
           </MDBox>
         ),
         [chart, height]
