@@ -19,7 +19,8 @@ import { useMemo } from "react";
 import PropTypes from "prop-types";
 
 // react-chartjs-2 components
-import { Line } from "react-chartjs-2";
+import 'chart.js/auto';
+import { Chart } from 'react-chartjs-2';
 
 // @mui material components
 import Card from "@mui/material/Card";
@@ -91,7 +92,7 @@ function DefaultLineChart({ icon, title, description, height, chart }) {
       {useMemo(
         () => (
           <MDBox height={height}>
-            <Line data={data} options={options} />
+            <Chart type="line" data={data} options={options} />
           </MDBox>
         ),
         [chart, height]
