@@ -19,7 +19,8 @@ import { useRef, useEffect, useState, useMemo } from "react";
 import PropTypes from "prop-types";
 
 // react-chartjs-2 components
-import { Line } from "react-chartjs-2";
+import 'chart.js/auto';
+import { Chart } from 'react-chartjs-2';
 
 // @mui material components
 import Card from "@mui/material/Card";
@@ -158,7 +159,7 @@ function MixedChart({ icon, title, description, height, chart }) {
       {useMemo(
         () => (
           <MDBox ref={chartRef} sx={{ height }}>
-            <Line data={data} options={options} />
+            <Chart type="line" data={data} options={options} />
           </MDBox>
         ),
         [chartData, height]
